Service_System/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
contourpy==1.3.3
crashtest==0.4.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
distlib==0.3.9
distro==1.9.0
Django==5.2.1
djangorestframework==3.16.1
dulwich==0.22.8
elastic-transport==8.17.1
elasticsearch==8.17.2
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
Flask==3.1.0
fonttools==4.61.1
frozenlist==1.5.0
fsspec==2025.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
installer==0.7.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.6.0
kiwisolver==1.4.9
langchain==0.3.22
langchain-community==0.3.20
langchain-core==0.3.50
langchain-ollama==0.3.0
langchain-openai==0.3.12
langchain-text-splitters==0.3.7
langgraph==0.3.25
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.24
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.8
matplotlib-inline==0.1.7
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.3.2
mypy-extensions==1.0.0
networkx==3.6.1
numpy==2.2.4
ollama==0.4.7
openai==1.70.0
opencv-python==4.12.0.88
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.3.3
parso==0.8.4
pbs-installer==2025.3.17
pexpect==4.9.0
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.7
poetry==2.1.2
poetry-core==2.1.2
polars==1.36.1
polars-runtime-32==1.36.1
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tomlkit==0.13.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
traitlets==5.14.3
trove-classifiers==2025.3.19.19
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.3
ultralytics==8.3.239
ultralytics-thop==2.0.18
urllib3==2.3.0
uvicorn==0.34.3
virtualenv==20.30.0
wcwidth==0.2.13
Werkzeug==3.1.3
xattr==1.1.4
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
